deleted old files new proj time
also create a new django start site and imma try to implement login  features thus needing forms and models and needing them interconnected, just did a lot of cybersec stuff so now i understand djangos design principles better also did a lot of reading on what the principles where and why and how to implement them. thus not i got a better understanding of them, also did a lot of research into http, and different protocols and readings about some vulnerabilities and how to secure my site better.

now i just got a better understanding of django in general and of backend development in general.

and of some of the features in django. that are implemented but i didnt know why or how they worked like cors and cross site scripting