okay getting further now trying to display the html forms correctly
and i just learned about widgets and that they render the html.

so forms have a way of being rendered and the widgets predefined those and forms just handle input logic and form validation

cool

and we are using the modelform

for the modelform

the modelform uses the model to create a form from. and thuse we have to specify the fields.

and we put those in a class Meta

not sure what that does tbh and the docs are difficult