reseted now following some tuturials on the auth system
there are a lot of things in the auth system also explored how the http and request thingies work in django and now i get it better able to understand the forms better that im creating rn,

also now import the login feature from django.contrib.auth

crazy and the logout feature is also there.

now im trying to register a person.

i have learned that you can add other fields to the default user fields. and switch that up.

also i know that you are able to validate if an user is logged in or if they have the right permissions. and that is connected to the database.

thus it must be possible to make a feature like that for my own added fields.

like a binary that can be true or false for permissions

and then create a control flow of only being able to receive the right site if that is set to one or to "true"

thus being able to create my own perms.

which is nice.