now trying out leaflet.js and update todo
im doing this because the google maps api is quuite expensive. and leaflet is self hosted.

also leaflet does exactly ythe same thing and is based on openmap.

openmap is a really good free use api. only have to give them credit, which is really nice and you can build ontop of it yourself.

that is what leaflet is.

pretty great no idea how to use openmap yet tho.

but i really like the leaflet integration and am curios to see how it is gonna work