rn making a protorype of django static files and the leaflet js.
dunno where im supposed to use folium,
just looked it up it also does have a geocoder plugin
setting up static files rn to see how it works havent used folium yet because i disliek the way how you import the map and pass it in the view and there create the coordinates and it generetes it all there it works but i kind of dislike it.

it makes in my head more sense to import it in the html and pass it in there but i think it does the same but i also disliek how i have to style it then and it seems nonsensical the structure becomes odd. and everything gets loaded from a different place but now its all in the same spot and i would still have to setup the css in the same way anyway but now i just have less control over it