the centre map div was broken if i posted the dropdown
okay so the for each only works on arrays and that broke the centre div thing so i fixed that.

by converting the json to an array for some reason that works the for each only works on arrays. and damn the converting of file between js and python is hell a lil atleast.